c4h_services/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "c4h_services"
version = "0.1.0"
authors = [
  { name="Jim" },
]
description = "Service implementations for c4h_agents"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "c4h_agents",
    "prefect>=2.14.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0", 
    "structlog>=23.2.0"
]

[project.optional-dependencies]
test = ["pytest>=7.0.0"]

[tool.hatch.build]
include = ["src/**/*.py"]

[tool.hatch.build.targets.wheel]
packages = ["src"]